requirements.txt
python-telegram-bot
tinydb
requests
bs4
pyquery
beautifulsoup4
python-dotenv
telethon
pandas
lxml
pyimgur
selenium
webdriver_manager
aiogram